creating a general wrapper for all of things in the project
