+++ src/components/Layout/Colors.module.scss
@@ -1,0 +1,5 @@
+$base-color: #FFCD1B;
+$black-color: #202020;
+$gray-color: #606060;
+$background-color: #e5e5e5;
+$white-color: #ffffff;
